pyproject.toml
[build-system]
requires = ["maturin>=1.4,<2.0"]
build-backend = "maturin"

[project]
name = "robotframework-javagui"
version = "0.3.0"
description = "High-performance Robot Framework library for Java GUI automation (Swing, SWT, Eclipse RCP)"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.8.1"
authors = [
    { name = "robotframework-javagui contributors" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Robot Framework",
    "Framework :: Robot Framework :: Library",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
]
keywords = ["robotframework", "swing", "swt", "rcp", "eclipse", "java", "automation", "testing", "gui"]
dependencies = [
    "docutils>=0.20.1",
    "robotframework>=4.0",
    "robotframework-assertion-engine>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/robotframework/robotframework-javagui"
Documentation = "https://robotframework-javagui.readthedocs.io"
Repository = "https://github.com/robotframework/robotframework-javagui"
Issues = "https://github.com/robotframework/robotframework-javagui/issues"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "JavaGui._core"
include = ["JavaGui/jars/*.jar"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "D", "UP", "S", "B", "A", "C4", "T20"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests/python"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "invoke>=2.2.1",
    "mypy>=1.14.1",
    "maturin>=1.4,<2.0",
    "robotframework-robocop>=5.5.0",
    "ruff>=0.14.11",
]
